Integration init

+++ src/main/scripts/resetDB.sql
@@ -5,7 +5,7 @@ CREATE TABLE products (
     id serial NOT NULL,
     name varchar(100),
     description varchar(250),
-    defaultPrice integer,
+    defaultPrice float,
     defaultCurrency varchar(250),
     category_id integer,
     supplier_id integer
